Utilize Client Serials instead of passing raw indexes.
There were a few bugs that were fixed with this refactoring as well, mainly IsClientInGame erroring out when the index is 0.
The main bug that was fixed is a client makes a query against a target, the target disconnects, and you're getting messages back from a potentially different target. This along with the client themselves disconnecting, someone else joins, and then they get the output of the command.

There may be some definite fallout from this, it was mostly manual labour.